requirements.txt
Flask==2.1.3
openai-whisper==20230124
pytest==7.2.1
redis==4.4.2
rq==1.12.0
python-dotenv==0.21.1
gunicorn==20.1.0
sentry-sdk==1.14.0
blinker==1.5
Werkzeug==2.0.3
Unidecode==1.3.6
requests==2.28.2